328 Downloaded widget component for captcha and moved NotFound to /Pages
